refac : action refacotring
