+++ src/actions/stock/stock.ts
@@ -31,20 +31,6 @@ async function getInvestors(stockCode: string) {
 
   return response.data
 }
-async function tradeStock(trade: string, tradeMoney: TradeType) {
-  const token = await getAccessToken()
-  const response = await PostAPI(`/trade/${trade}`, tradeMoney, token)
-  return response
-}
-async function tradeReservation(trade: string, tradeMoney: TradeType) {
-  const token = await getAccessToken()
-  const response = await PostAPI(
-    `/trade/reservation/${trade}`,
-    tradeMoney,
-    token,
-  )
-  return response
-}
 
 async function getStockData(stockCode: string, when: string) {
   const response = await GetAPI(`/stockitem/chart/${stockCode}/${when}`)
@@ -59,8 +45,6 @@ export {
   getStockName,
   getChartData,
   getInvestors,
-  tradeStock,
-  tradeReservation,
   getStockData,
   getStaticStockPrice,
   getStaticStockAskPrice,
